fix: Use the latest beta of the hyprland crate
Prior to this commit when performing a regular debug `cargo build`
everything would work just fine, but performing a `cargo build
--release` or a `cargo install --path . --locked` would fail.

It appears that cargo was pulling the beta version of the hyprland crate
when building in debug mode, but the alpha version when building in
release mode.

Since it's simple enough to upgrade to the beta version of the create,
let's go ahead and do that.